fastapi_backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
asyncpg>=0.29.0
playwright>=1.40.0
pillow>=10.1.0
python-slugify>=8.0.1
aiofiles>=23.2.1